# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent Hub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  password: string;
}

export interface LoginResponse {
  success: boolean;
}

export interface AuthStatus {
  authenticated: boolean;
}

export interface Agent {
  id: number;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  backstory?: string | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  coreValues?: string | null;
  /** @nullable */
  expertiseAreas?: string | null;
  /** @nullable */
  communicationStyle?: string | null;
  /** @nullable */
  emotionalIntelligence?: string | null;
  language: string;
  isActive: boolean;
  connectionsCount: number;
  /** @nullable */
  lastActivity?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAgentBody {
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  backstory?: string | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  coreValues?: string | null;
  /** @nullable */
  expertiseAreas?: string | null;
  /** @nullable */
  communicationStyle?: string | null;
  /** @nullable */
  emotionalIntelligence?: string | null;
  language?: string;
}

export interface UpdateAgentBody {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  backstory?: string | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  coreValues?: string | null;
  /** @nullable */
  expertiseAreas?: string | null;
  /** @nullable */
  communicationStyle?: string | null;
  /** @nullable */
  emotionalIntelligence?: string | null;
  language?: string;
  isActive?: boolean;
}

export interface KnowledgeEntry {
  id: number;
  agentId: number;
  type: string;
  /** @nullable */
  title?: string | null;
  content: string;
  /** @nullable */
  sourceUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type AddKnowledgeBodyType =
  (typeof AddKnowledgeBodyType)[keyof typeof AddKnowledgeBodyType];

export const AddKnowledgeBodyType = {
  text: "text",
  document: "document",
  url: "url",
} as const;

export interface AddKnowledgeBody {
  type: AddKnowledgeBodyType;
  /** @nullable */
  title?: string | null;
  content: string;
  /** @nullable */
  sourceUrl?: string | null;
}

export interface UpdateKnowledgeBody {
  /** @nullable */
  title?: string | null;
  content?: string;
  /** @nullable */
  sourceUrl?: string | null;
}

export interface Instruction {
  id: number;
  agentId: number;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface AddInstructionBody {
  content: string;
}

export interface UpdateInstructionBody {
  content?: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  agentId: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface SendChatMessageBody {
  message: string;
}

export interface Connection {
  id: number;
  agentId: number;
  appName: string;
  apiKey: string;
  /** @nullable */
  lastUsed?: string | null;
  requestCount: number;
  createdAt: string;
}

export interface CreateConnectionBody {
  appName: string;
}

export interface ActivityEntry {
  id: number;
  agentId: number;
  /** @nullable */
  connectionId?: number | null;
  /** @nullable */
  appName?: string | null;
  userMessage: string;
  agentResponse: string;
  createdAt: string;
}

export type ConversationTurnRole =
  (typeof ConversationTurnRole)[keyof typeof ConversationTurnRole];

export const ConversationTurnRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ConversationTurn {
  role: ConversationTurnRole;
  content: string;
}

export interface PublicChatBody {
  apiKey: string;
  message: string;
  conversationHistory?: ConversationTurn[];
  /** @nullable */
  userId?: string | null;
}

export interface PublicChatResponse {
  response: string;
  agentName: string;
}