# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Agent Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /agents/{id}/toggle-status:
    post:
      operationId: toggleAgentStatus
      tags: [agents]
      summary: Toggle agent active/inactive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
...
  /agents/{agentId}/chat:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get owner chat history
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send message to agent (owner chat, streaming)
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageBody"
      responses:
        "200":
          description: Streaming SSE response
    delete:
      operationId: clearChatHistory
      tags: [chat]
      summary: Clear chat history
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cleared
  /agents/{agentId}/connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List connections for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      operationId: createConnection
      tags: [connections]
      summary: Create a connection (generate API key)
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionBody"
      responses:
        "201":
          description: Connection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
  /agents/{agentId}/connections/{id}:
    delete:
      operationId: deleteConnection
      tags: [connections]
      summary: Revoke connection
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /agents/{agentId}/activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: List activity log for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEntry"
...
components:
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        backstory:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        coreValues:
          type: ["string", "null"]
        expertiseAreas:
          type: ["string", "null"]
        communicationStyle:
          type: ["string", "null"]
        emotionalIntelligence:
          type: ["string", "null"]
        language:
          type: string
        isActive:
          type: boolean
        connectionsCount:
          type: integer
        lastActivity:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - language
        - isActive
        - connectionsCount
        - createdAt
        - updatedAt
    CreateAgentBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        backstory:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        coreValues:
          type: ["string", "null"]
        expertiseAreas:
          type: ["string", "null"]
        communicationStyle:
          type: ["string", "null"]
        emotionalIntelligence:
          type: ["string", "null"]
        language:
          type: string
      required:
        - name
    UpdateAgentBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        backstory:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        coreValues:
          type: ["string", "null"]
        expertiseAreas:
          type: ["string", "null"]
        communicationStyle:
          type: ["string", "null"]
        emotionalIntelligence:
          type: ["string", "null"]
        language:
          type: string
        isActive:
          type: boolean
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - agentId
        - role
        - content
        - createdAt
    SendChatMessageBody:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Connection:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: integer
        appName:
          type: string
        apiKey:
          type: string
        lastUsed:
          type: ["string", "null"]
        requestCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - agentId
        - appName
        - apiKey
        - requestCount
        - createdAt
    CreateConnectionBody:
      type: object
      properties:
        appName:
          type: string
      required:
        - appName
    ActivityEntry:
      type: object
      properties:
        id:
          type: integer
        agentId:
          type: integer
        connectionId:
          type: ["integer", "null"]
        appName:
          type: ["string", "null"]
        userMessage:
          type: string
        agentResponse:
          type: string
        createdAt:
          type: string
      required:
        - id
        - agentId
        - userMessage
        - agentResponse
        - createdAt
...