# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEntry {
  id: number;
  agentId: number;
  /** @nullable */
  connectionId?: number | null;
  /** @nullable */
  appName?: string | null;
  userMessage: string;
  agentResponse: string;
  createdAt: string;
}