# lib/api-zod/src/generated/types/connection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Connection {
  id: number;
  agentId: number;
  appName: string;
  apiKey: string;
  /** @nullable */
  lastUsed?: string | null;
  requestCount: number;
  createdAt: string;
}