# lib/api-zod/src/generated/types/createAgentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAgentBody {
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  backstory?: string | null;
  /** @nullable */
  personality?: string | null;
  /** @nullable */
  coreValues?: string | null;
  /** @nullable */
  expertiseAreas?: string | null;
  /** @nullable */
  communicationStyle?: string | null;
  /** @nullable */
  emotionalIntelligence?: string | null;
  language?: string;
}